pyproject.toml
[project]
name = "ais-dev2il-smoothie-shop"
version = "0.1.0"
description = "The Smoothie Shop"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
    "python-logging-loki>=0.3.1",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-httpx>=0.49b0",
    "opentelemetry-exporter-otlp>=1.28.0",
]
